antlir oss: workflow to build and upload ba
Summary:
Create a GitHub Actions workflow that builds and uploads a new version of the
stable build appliance image in S3 whenever the appliance BUCK file changes (a
proxy for deps, since it doesn't have any feature deps, just the compiler).

Note that commits that simply update the revision (such as the automatic PRs)
do not trigger this rebuild flow, since it is in a separate file.

Reviewed By: zeroxoneb

Differential Revision: D25726235

fbshipit-source-id: 9af54a5d2b632bbd419dbe4f619d913189153525